refactor to make smallest change possible
