+++ plugins/yaml/src/org/jetbrains/yaml/YAMLLanguageCodeStyleSettingsProvider.java
@@ -1,7 +1,9 @@
 // Copyright 2000-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package org.jetbrains.yaml;
 
-import com.intellij.application.options.*;
+import com.intellij.application.options.CodeStyleAbstractPanel;
+import com.intellij.application.options.IndentOptionsEditor;
+import com.intellij.application.options.SmartIndentOptionsEditor;
 import com.intellij.lang.Language;
 import com.intellij.psi.codeStyle.*;
 import org.jetbrains.annotations.NotNull;
@@ -13,6 +15,9 @@ import javax.swing.*;
 import static com.intellij.psi.codeStyle.CodeStyleSettingsCustomizableOptions.getInstance;
 
 public class YAMLLanguageCodeStyleSettingsProvider extends LanguageCodeStyleSettingsProvider {
+
+  private static final YAMLCodeStyleSettingsProvider CODE_STYLE_SETTINGS_PROVIDER = new YAMLCodeStyleSettingsProvider();
+
   private static class Holder {
     private static final int[] ALIGN_VALUES = new int[]{
       YAMLCodeStyleSettings.DO_NOT_ALIGN,
@@ -27,39 +32,21 @@ public class YAMLLanguageCodeStyleSettingsProvider extends LanguageCodeStyleSett
     };
   }
 
-  @NotNull
   @Override
-  public CodeStyleConfigurable createConfigurable(@NotNull final CodeStyleSettings settings, @NotNull final CodeStyleSettings originalSettings) {
-    return new CodeStyleAbstractConfigurable(settings, originalSettings, YAMLLanguage.INSTANCE.getDisplayName()) {
-      @Override
-      protected CodeStyleAbstractPanel createPanel(final CodeStyleSettings settings) {
-        final CodeStyleSettings currentSettings = getCurrentSettings();
-        return new TabbedLanguageCodeStylePanel(YAMLLanguage.INSTANCE, currentSettings, settings) {
-          @Override
-          protected void initTabs(final CodeStyleSettings settings) {
-            addIndentOptionsTab(settings);
-            addSpacesTab(settings);
-            addWrappingAndBracesTab(settings);
-          }
-        };
-      }
-
-      @Override
-      public String getHelpTopic() {
-        return "reference.settingsdialog.codestyle.yaml";
-      }
-    };
+  public @NotNull CodeStyleConfigurable createConfigurable(@NotNull CodeStyleSettings baseSettings,
+                                                           @NotNull CodeStyleSettings modelSettings) {
+    return CODE_STYLE_SETTINGS_PROVIDER.createConfigurable(baseSettings, modelSettings);
   }
 
   @Override
   public String getConfigurableDisplayName() {
-    return YAMLLanguage.INSTANCE.getDisplayName();
+    return CODE_STYLE_SETTINGS_PROVIDER.getConfigurableDisplayName();
   }
 
   @Nullable
   @Override
   public CustomCodeStyleSettings createCustomSettings(CodeStyleSettings settings) {
-    return new YAMLCodeStyleSettings(settings);
+    return CODE_STYLE_SETTINGS_PROVIDER.createCustomSettings(settings);
   }
 
   @Override
